# swagger.json
{
    "swagger": "2.0",
    "info": {
      "description": "Swagger UI for the Chime API",
      "version": "1.0.0",
      "title": "Swagger - Chime"
    },
    "host": "chime-api-v2.herokuapp.com",
    "basePath": "",
...
    "schemes": [
      "https",
      "http"
    ],
    "paths": {
...
      "/profile": {
        "get": {
          "tags": [
            "profile"
          ],
          "summary": "Get profile",
          "description": "Returns profile of given username",
          "operationId": "getProfile",
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "username",
              "in": "query",
              "description": "Username of profile that needs to be fetched",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "Success",
              "schema": {
                "$ref": "#/definitions/Profile"
              }
            },
            "400": {
              "description": "Missing parameters"
            },
            "404": {
              "description": "Account not found"
            },
            "500": {
              "description": "Internal server error"
            }
          }
        },
        "put": {
          "tags": [
            "profile"
          ],
          "summary": "Update profile",
          "description": "Updates profile for a user",
          "operationId": "updateProfile",
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "body",
              "in": "body",
              "description": "New profile details and authentication details",
              "required": true,
              "schema": {
                "$ref": "#/definitions/ProfileAuth"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success",
              "schema": {
                "$ref": "#/definitions/Profile"
              }
            },
            "400": {
              "description": "Missing parameters"
            },
            "403": {
              "description": "Authentication failed"
            },
            "500": {
              "description": "Internal server error"
            }
          }
        }
      },
...
    },
    "definitions": {
...
      "Profile": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "rsvpmx"
          },
          "first": {
            "type": "string",
            "example": "John"
          },
          "last": {
            "type": "string",
            "example": "Smith"
          },
          "occupation": {
            "type": "string",
            "example": "Student"
          },
          "picture": {
            "type": "string",
            "example": "https://i.imgur.com/FSgbIi4.png"
          },
          "about": {
            "type": "string",
            "example": "I like food."
          },
          "birthday": {
            "type": "string",
            "example": "12/12/1997"
          },
          "location": {
            "type": "string",
            "example": "Auckland"
          }
        }
      },
...
      "ProfileAuth": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "rsvpmx"
          },
          "password": {
            "type": "string",
            "example": "secret123"
          },
          "occupation": {
            "type": "string",
            "example": "Student"
          },
          "picture": {
            "type": "string",
            "example": "--needs to be a base64 string--"
          },
          "about": {
            "type": "string",
            "example": "I like food."
          },
          "birthday": {
            "type": "string",
            "example": "12/12/1997"
          },
          "location": {
            "type": "string",
            "example": "Auckland"
          }
        }
      },
...
    }
  }